added implementation of WeatherDataProvider to MetClient

+++ src/main/java/com/example/weatheroptimizer/clients/model/Humidity.java
@@ -1,4 +1,4 @@
 package com.example.weatheroptimizer.clients.model;
 
-public record Humidity(Integer value, String unit) {
+public record Humidity(Double value, String unit) {
 }
